renamed base -> _base
